# backend/openapi/openapi.yaml
...
openapi: 3.1.0
info:
  title: United Pass Backend API
  description: |
    United Pass is a unified identity, account, organization, OAuth application,
    authorization, Provider synchronization and audit platform.

    This OpenAPI specification is the machine-readable contract. The human-readable
    detail lives in `frontend/docs/api-contracts.md`. Both must stay synchronized.

    Phase 0 scope: operational endpoints only.
    Phase 1 scope: session establishment, MFA, logout, current user and
    permission capabilities.
    Phase 2 scope: OAuth application and client management plane (scope
    catalog, application/client lifecycle, secret rotation, reauthentication).
    See `../docs/adr-0004.md` for the architecture decisions.

    Business API paths use the full `/api/v1/...` prefix explicitly rather than
    relying on a global server base path, so that root-level operational endpoints
    and versioned API endpoints can coexist accurately.
  version: 0.3.0
  contact:
    name: United Pass backend team
servers:
  - url: http://localhost:8080
    description: Local development
  - url: /
    description: Same-origin production
...
paths:
...
  /api/v1/auth/sessions:
    post:
      tags: [Session]
      summary: Establish a browser session
      description: |
        Authenticates the caller with a unified identifier (account name or email)
        and password, then establishes a browser session.

        On success the server issues two cookies:
          * `up_session` — HttpOnly, Secure, SameSite session cookie.
          * `up_csrf`   — CSRF token cookie readable by browser JavaScript.

        Both cookies are returned via `Set-Cookie`. The browser must not receive
        raw authentication-provider Refresh Tokens or long-lived tokens.

        When the account requires multi-factor authentication, the server responds
        `202` with an opaque `mfaToken` and the list of available methods instead
        of establishing the session. The MFA challenge is completed via
        `POST /api/v1/auth/sessions/mfa`.

        Authentication failures return generic errors that do not reveal whether
        the account, email or phone number exists. Passwords are never logged.

        Rate limited. Repeated failures return `429` with a `Retry-After` header.
      operationId: createSession
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
            examples:
              credentials:
                value:
                  identifier: zhixing.lin
                  password: user-entered-password
                  remember: true
              resumeAuthorization:
                value:
                  identifier: zhixing.lin
                  password: user-entered-password
                  remember: false
                  resumeRequestId: req_01ABCdefGHIjkl
      responses:
        "204":
          description: |
            Session established successfully. No response body. The
            `up_session` and `up_csrf` cookies are set.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/X-Request-ID"
            Set-Cookie:
              $ref: "#/components/headers/Set-Session-Cookies"
        "202":
          description: |
            Multi-factor authentication is required before the session can be
            established. The caller must complete the challenge at
            `POST /api/v1/auth/sessions/mfa`.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/X-Request-ID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MFARequiredResponse"
              example:
                status: mfa_required
                mfaToken: opaque_mfa_token
                availableMethods: [totp, passkey, recovery_code]
                expiresAt: "2026-08-05T12:05:30Z"
        "400":
          description: Malformed request body or missing required fields.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/X-Request-ID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: validation.failed
                  message: "请求参数有误，请检查后重试。"
                  requestId: req_0123456789abcdef
                  fieldErrors:
                    - field: identifier
                      message: "该字段不能为空。"
        "401":
          description: |
            Invalid credentials. The response is generic and does not reveal
            whether the identifier corresponds to an existing account.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/X-Request-ID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: auth.invalid_credentials
                  message: "账户名、邮箱或密码错误。"
                  requestId: req_0123456789abcdef
        "413":
          description: Request body exceeds the configured maximum size.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/X-Request-ID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: request.too_large
                  message: "请求体过大。"
                  requestId: req_0123456789abcdef
        "429":
          description: |
            Too many login attempts. The caller must wait before retrying.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/X-Request-ID"
            Retry-After:
              $ref: "#/components/headers/Retry-After"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: auth.rate_limited
                  message: "尝试次数过多，请稍后再试。"
                  requestId: req_0123456789abcdef
        "500":
          description: Unexpected server failure.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/X-Request-ID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: internal.unexpected
                  message: "服务暂时无法处理该请求。"
                  requestId: req_0123456789abcdef
...
  /api/v1/auth/session:
    delete:
      tags: [Session]
      summary: Revoke the current browser session (logout)
      description: |
        Revokes the server-side session bound to the `up_session` cookie and
        instructs the browser to discard both the `up_session` and `up_csrf`
        cookies by setting them with `Max-Age=0`.

        Requires a valid session cookie and a matching `X-CSRF-Token` header.
        The CSRF token must equal the value of the `up_csrf` cookie.
      operationId: deleteSession
      security:
        - SessionCookie: []
      parameters:
        - $ref: "#/components/parameters/X-CSRF-Token"
      responses:
        "204":
          description: |
            Session revoked successfully. No response body. The `up_session`
            and `up_csrf` cookies are cleared.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/X-Request-ID"
            Set-Cookie:
              $ref: "#/components/headers/Clear-Session-Cookies"
        "401":
          description: No valid session cookie was supplied.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/X-Request-ID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: auth.unauthenticated
                  message: "请先登录后再继续。"
                  requestId: req_0123456789abcdef
        "403":
          description: |
            CSRF validation failed. The `X-CSRF-Token` header is missing or does
            not match the `up_csrf` cookie.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/X-Request-ID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: csrf.validation_failed
                  message: "跨站请求验证失败，请刷新页面后重试。"
                  requestId: req_0123456789abcdef
        "500":
          description: Unexpected server failure.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/X-Request-ID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: internal.unexpected
                  message: "服务暂时无法处理该请求。"
                  requestId: req_0123456789abcdef
...
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: up_session
      description: |
        Browser session cookie (`up_session`) issued by United Pass. The cookie
        is HttpOnly, Secure and carries an explicit SameSite policy. The browser
        must not receive raw authentication-provider Refresh Tokens or
        long-lived tokens.
  headers:
    X-Request-ID:
      description: |
        The request ID used for log, error and audit correlation. The server
        accepts a valid upstream ID (alphanumeric, dash, underscore, colon;
        8-128 chars) or generates a cryptographically random `req_`-prefixed ID.
      schema:
        type: string
        pattern: "^[a-zA-Z0-9_:-]{8,128}$"
        example: req_0123456789abcdef0123456789abcdef
    Retry-After:
      description: |
        Seconds the client must wait before retrying a rate-limited request.
        Returned with `429` responses.
      schema:
        type: integer
        format: int64
        minimum: 1
        example: 30
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/ErrorBody"
...
    LoginRequest:
      type: object
      description: |
        Credentials used to establish a browser session. `identifier` accepts
        either an account name or an email address. Authentication failures
        return generic errors that do not reveal which value was incorrect.
      required: [identifier, password]
      additionalProperties: false
      properties:
        identifier:
          type: string
          minLength: 1
          description: Account name or email address used to identify the user.
          example: zhixing.lin
        password:
          type: string
          minLength: 1
          format: password
          description: User-entered password. Never logged or persisted in plaintext.
          writeOnly: true
        remember:
          type: boolean
          default: false
          description: |
            When true, the session cookie is issued with an extended lifetime
            suitable for a "remember me" experience. When false or omitted, the
            session uses the default browser-session lifetime.
        resumeRequestId:
          type: string
          description: |
            Opaque server-issued authorization transaction ID to resume after
            successful login. When present and valid, login redirects to the
            authorization resume flow instead of the default account or admin
            destination. Only an opaque server-issued ID is accepted — never a
            raw return URL.
          pattern: "^[a-zA-Z0-9_:-]{8,128}$"
          example: req_01ABCdefGHIjkl
...